Full overhaul with tests
This commit is an almost complete refactor of the previous module. It removes
some of the levers and switches with the repos and instead of using the name of
a mock config to figure out how it is laid out, this uses the parameters of a
mock config to determine its name. The entry point of init.pp now collects some
common params and hands them down to the other manifests if provided.