Add epel repo with only ccache included
Without the epel repo enabled on el mocks, the buildsys-build group will be
unavailable for installation. This commit adds the epel repo to the mock config
for el mocks so that the buildsys-build group can be found and the mock
created. Only the ccache package is enabled, so nothing else will be brought
in.