Add ccache to pe-mock, add header to FOSS template
This commit adds the epel repo with just the ccache package to the pe-mock
config. It also updates the chroot_setup_cmd to install the buildsys-build
group which is available once the epel repo is added. It also adds a header to
the FOSS mock config to identify it as a puppet template.