Add dependency on puppetlabs_yum and use it
This removes the logic in the repos manifest and instead uses a module which
already implements the same logic. It adds the dependency in the modulefile,
.fixtures.yml. It also adds a new param to the main class which can be used to
decide whether or not to include the pl repos.
This also moves the include on epel into init.pp and updates the previous
package dependency from rpmbuilder::repos to epel.