Change groupinstall to install for setup mock command
Ah, el4, when yum was new and up2date was terrible. In this case yum
hadn't yet learned about groups. Therefore, metapackages were built that
contained nothing but had requirements on items to be pulled in for
building. That was already defined in the legacy-mock-config template
but the setup/install command was still referencing groupinstall. This
commit fixes that. In other words, s/groupinstall/install