(RE-812) Update mock template with temporary band-aid
Since RHEL7 isn't really a released, baked, finished thing yet,
we have to reference the beta. This has a few caveats:
- x86_64 only
- None of the existing mock templates quite work

For the time being, support is being provided via a 'bandaid',
which simply declares a file resource for the mock config if a
RHEL7 host is to be built. This scaffold/band-aid/kludge will
be rolled back/removed/killed with fire when RHEL7 gets its wish
and the Blue Fairy makes it a real release.