(maint) Pin EL-{5,6} mocks to the ".0" release
In theory, the initial release of EL in each family should
be forward-compatible through the lifecycle. The reverse is
not guaranteed to be true.

This commit pins the EL 5 and 6 mocks against the ".0" release
of each EL family to avoid issues with, for example, a
later version of a library adding (or removing) symbols that make
software compiled against it incompatible with earlier versions
of the library.