(RE-765) Add EOS4 mocks to pe_mockset
This commit adds a template for EOS mocks to the module and adds the correct
line in pe_mockset to add the mock for all of our PE versions. The mock config
is based on fedora and the set of rpms that make up EOS4. Because
fedora-release depends on Eos-release and that is currently unavailable, this
mock config currently uses the verbose install rather than a groupinstall.