(RE-2158) Add build-tools param to rpmbuilder module
This commit adds a top level parameter to the rpmbuilder module to
enable the pl-build-tools repo, an internal repository to Puppet Labs.
It defaults to false to ensure backward compatibility. That parameter is
directly looked up in both the el7 and pl mock templates. This is to
avoid passing the parameter down through 5 levels of resource
indirection.