(maint) Repoint RHEL7 repos to internal Puppet Labs resources
This change is a short-term fix until CentOS 7 is released. For the
time being, anyone outside Puppet Labs who uses this module is
recieving a 404 error. After this patch, they will recieve an
NXDOMAIN error, which is different but no less broken than current
behavior. We should be able to reverse this with the release of
CentOS 7, as well as removing the RHEL7 band-aid configurations.