(RE-2465) Remove mirrorlist declarations from PE mock configs
Because all of the PE EL oses are mirrored internally, we shouldn't need
to configure a mirrorlist for them. This combined with setting a proxy,
causes bad things to happen if the repodata between the two doesn't
agree (for say relative pathing). This commit removes those declarations
from the el and sles templates.