(maint) Enable build-tools repo on all EL dists for PE
Previously the build-tools repo was only enabled on el 5 and 6. With the
addition of the legacy mock to the template, el4 is also required. And
el7 has a build-tools repo available as well. This commit updates that
logic to just check for el and add the repo if on el.