(RE-2466) Remove legacy mock defined types
Previous commits removed the need for a legacy mock in templates by
adding a few conditional stanzas. This commit makes a similar change to
the puppet code, removing the pe_legacy_mock and pe_legacy_mockset
defined types. It also updates the pe_mocks and pe_mockset to include
el4 mocks for PE.