(re-1897) Add sles 10 mocks to rpmbuilder
This adds SLES 10 to the PE mocksets. Since we will typically
build against SLES sp1, but we need to build against sp4 for SLES 10
(as it is the only supported version currently) this adds a parameter for
the SLES mocks that defaults to 'sp1' and is overridden for the SLES 10
mocks.