(maint) Update release comparisons in pe-mock template to strings
Previously the @release variable was compared to the integers 4 and 7,
which always returns false because 4 and 7 will be "4" and "7", which
are strings. This commit updates the pe-el mock template to do string
comparisons instead of integer comparisons. It also updates the erb
conditionals to not generate extra whitespace for easier to read mock
configs.