(bugfix) Define %dist macro for all el versions
On el7, `%dist` was being evaluated as `el7.centos`, which is not what we
want. Because of this, we need to override the definition of `%dist` on
el7 as well. Since there isn't any reason to just do this everywhere,
(i.e. not to skip el6), this also removes the conditional assignment of
`%dist`