(maint) Disable proxy for internal mocks
We run into problems using a proxy for epel in that repodata is being
cached, despite attempts to disable this on the proxy. This commit
removes the proxy from the pe mock configs so that builds can progress
without error. We only pull in ccache and rpm groups from epel, so the
added cost should be negligible.