(maint) Add autoconf to SLES mock stubs
BuildRequires are apparently sort-of unreliable on SLES which is
causing the HTTPD package to disregard them when building HTTPD.
Subsequently, autoconf doesn't get installed so the build fails.
This adds in the troublesome package, which will probably be of
benefit somewhere up the road.

This commit also alphabetizes the packages installed into the
SLES mock stubs.