(maint) Add pl-build-tools to pe mock config
Since we have started building out the compiled software toolchain, we
have decided to ship necessary dependency packages to
pl-build-tools.delivery.puppetlabs.net. This was mainly because we
couldn't decide if we wanted to make those packages publically available
or not. Because of this, we need to ensure that repo is available to the
mocks we are using to build those compiled software packages (like
cfacter). We've already done this for the FOSS mocks, but we need to do
this for the pe mocks as well. This is specifically required for
building el4 packages, since we currently do not have el4 mocks for
FOSS, and are instead using the PE el4 mocks to build these packages.
This commit will make this the case on all our builders.