(RE-4290) Pin mock to a known good version
Rather than always pull in the latest version of mock available, we
should always be using a known good version. This has become a necessity
because we aren't doing a good job with our unit tests. With the upgrade
to mock, our spec tests were actually seeing through the mock to the
base system, which was causing failures. Since we're moving to vanagon,
this isn't going to matter much longer, hence why we're just pinning
here, rather than fixing the testing.