(RE-7420) Add sles logic to mock config
This commit adds in logic to enable repos for the sles mocks. The
problem is that sles doesn't provide publicly available mirrors. We have
set up an internal mirror, but this isn't going to work outside of
Puppet, Inc. infrastructure. This seems like the cleanest workaround to
this issue, rather than linking a mock with a defined version of PE.
This allows us to create sles mocks for the public which should just not
access any repos and also not fail because it has enabled inaccessible
repos. That would be an issue if we were to link the pl-sles mocks to a
pupent-sles mock as the pupent-sles mocks enable the internal sles repos
with no fallback logic. That, and we no longer have to decide which PE
version the foss mocks will be tied to. This frees us up from the worry
of someone deleting on obsolete PE mock config, which happened to be the
one we were dependent on with sles.