(RE-8152) Use internal mirrors when building
Prior to this commit, we had been setting all mocks to install from
publicly available and externally maintained mirrors. This is fine,
and it is requires as this module is publicly available. However, this
does sometimes prove to be a pain point sometimes. It is not best
practice to rely on external mirrors always being up and available. This
commit updates this module to allow a user to pass in a variable
$pl_internal. This is only meant to be used inside the Puppet, Inc. VPN.
This will set the baseurl for the system repos to be set to the mirror
that we maintain internally. This will help insulate us from future
changes such as platform EOL status.